Fix find command related bugs and refactor code for optional fields
Optional person attribute classes, such as risk and meeting date, are
required to be able to take and store "empty" inputs in the case that
the user leaves that field empty. The current implementation stores
empty values as null, and when these fields are needed, the consumer are
then required to check for null values themselves, leading to code
duplication and leaving the code prone to errors, as it is difficult to
ensure that every consumer does a check for nullary values. This
behaviour is hence undesirable, as it violates both the Don't Repeat
Yourself (DRY) principle, and breaks the abstraction barrier.

This check for nullary values can be extracted out into each of the
optional personal attribute classes by using the java `Optional`
abstraction. This allows the nullary check to be performed within the
person attribute field class itself, without its inner value being
leaker out to its consumers.