Fix error message to display message specific to the index
Before, entering a negative integer will result in the command usage being displayed.
After this change, entering a negative integer will instead result in the message
"Index is not a non-zero unsigned integer." being displayed.
Entering a positive integer that is out of bounds of the client list will display the more
specific message, "The client index provided could not be found in the list of clients."