Fix a bug where an edited meeting will conflict with itself
Before, if we tried editing a meeting, e.g., from 1300 to 1400 and make its start
time 1230, it will give an error that there is a conflicting meeting because it
conflicts with the current version of itself.
After this fix, the meetings list is only checked for conflicts with meetings other
than the one being edited.