Require whitespace after flag when tokenizing args
in AB3, the flag was in the format "<flag>/", thus the ArgumentTokenizer
only checked for a whitespace before the flag to recognise it as that
flag. this implementation was not changed when we changed the flag format to
"-<flag>". this continues to work, as long as we do not have flags that
had the same prefix (e.g. "-p" and "-ph"), in which case both "-ph" and "-p"
would be recognised as the "-p" prefix.

however, we recently standardised the flags in an effort to make it
consistent in all PRs. this resulted in the creation of the "-p" flag,
which is a prefix of the existing "-ph" flag, causing the above issue.

lets require a whitespace after the flag as well when tokenizing
arguments.