Make changes to UG
For the getting started section, the 2 example pictures for the 2 Ui
layouts are on top of one another, making it hard to compare.

Let's put it side by side with headers so readers can clearly see and
compare the 2 layouts.

The viewing tips section is at near the end of the guide. It is one of
our defining features and should be more front and centre.

Let's move it up right below the list command section so after users add
and list their applications, they will know how to view the tips right
after.

The light bulb tutorial in the viewing tips section has example images
on top of one another, making it hard for users to group the examples
together by making them scroll unnecessarily.

Let's put it side by side with headers so users can easily view both
example images at a glance.

Additionally, some headers and texts were changed to be more appropriate
and user friendly.