Fix user guide
Used more personal language to make GuestBook more friendly to read.

Update User Guide command examples so that it can be copied and pasted seamlessly to be used in GuestBook (closes #156)

Clarified that finding dates only work through the start and end dates (closes #172)

Clarified that finding a date range can return other results as well (closes #175)

Clarified that the room numbers must be unique in GuestBook irregardless of dates of stay of a guest (closes #172).