Amend bug for edit module feature
Initially, when the edit module command is called, all exams with
the module to be edited has their module field changed to the
edited module. Also, all tasks with the module to be edited has
their module field changed to the edited module. However the
original links between these tasks and exams were not preserved.

To ensure these links are preserved, the DistinctTaskList and
DistinctExamList classes were augmented with methods to check
these links.