Make the unavailable date and home visit to display in order
It is to fix bug as the edit, deassign operation depends on the index
given by user and also on the sorted home visit and unavailable date
list. Hence, if the display list is not sorted, then the user might
give index that does not corresponds to what they intend to change.

Hence, this is a bug that need to be fixed.